apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
requests==2.32.5
python-dotenv==1.0.1
